restructure the mocks to make more sense
